Pull the tests out of lib.rs into their own file
Better code organization and removes a level of indentation
